+++ src/client/TextRenderer.cpp
@@ -77,7 +77,7 @@ void TextRenderer::Initialize()
         Character character = {texture,
                                glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                                glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                               face->glyph->advance.x};
+                               (GLuint)face->glyph->advance.x};
         s_characters.insert(std::pair<GLchar, Character>(c, character));
     }
     glBindTexture(GL_TEXTURE_2D, 0);
